+++ src/index.ts
@@ -1,31 +1,17 @@
 import cors from 'cors';
 import express from 'express';
-import session from 'express-session';
 import 'reflect-metadata';
 import AppDataSource from './config/dataSource';
 import './config/env';
 import router from './controller/router';
 import errorHandler from './util/errorHandler';
+import sessionMiddleware from './middleware/session';
 
 const PORT = Number(process.env.PORT) || 3000;
 
-const SESSION_SECRET = String(process.env.SESSION_SECRET);
-
 const app = express();
-
-declare module 'express-session' {
-  export interface SessionData {
-    user: { [key: string]: any };
-  }
-}
-
-// 세션 미들웨어 설정 (gpt 행님)
-app.use(session({
-  secret: SESSION_SECRET, // 세션 암호화에 사용되는 비밀키
-  resave: false,
-  saveUninitialized: true,
-  cookie: { secure: false }, // HTTPS가 아닌 환경에서도 사용 가능하도록 설정
-}));
+//sessionMiddleware 별도 관리
+app.use(sessionMiddleware);
 
 AppDataSource.initialize().then(() => console.log('DATABASE is connected!'));
